+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "app",
   "version": "1.0.0",
-  "main": "index.ts",
+  "main": "expo-router/entry",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
@@ -14,9 +14,13 @@
     "expo-font": "^13.0.3",
     "expo-status-bar": "~2.0.1",
     "react": "18.3.1",
-    "react-native": "0.76.6",
-    "react-native-safe-area-context": "^5.2.0",
-    "react-native-web": "~0.19.13"
+    "react-native": "0.76.7",
+    "react-native-safe-area-context": "4.12.0",
+    "react-native-web": "~0.19.13",
+    "expo-router": "~4.0.17",
+    "react-native-screens": "~4.4.0",
+    "expo-linking": "~7.0.5",
+    "expo-constants": "~17.0.7"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
