+++ src/components/MainText.tsx
